Udemy/Premium Telegram Bot in Python/requirements.txt
python-telegram-bot==21.10
python-dotenv==1.0.1
nest-asyncio==1.5.5
boto3==1.36.8